refactor(ui): move and refactor ApiKey interfaces
