+++ ui/src/interfaces/IUserLogin.ts
@@ -29,39 +29,3 @@ export interface MfaReset {
   main_email_code: string,
   recovery_email_code: string,
 }
-
-enum ApiKeyExpireList {
-  NUMBER_30 = 30,
-  NUMBER_60 = 60,
-  NUMBER_90 = 90,
-  NUMBER_365 = 365,
-  NUMBER_MINUS_1 = -1,
-}
-export interface ApiKeyValidation {
-  tenant: string,
-  name: string,
-  role: string,
-  expires_at: ApiKeyExpireList,
-}
-
-export interface ApiKeyEdit {
-  key: string,
-  name: string,
-  role: string,
-}
-
-export interface ApiKey {
-  name: string,
-  expires_at: string,
-}
-
-export interface ApiKeyRemove {
-  key: string,
-}
-export interface ApiKeyGetValidation {
-  tenant: string,
-  page : number,
-  perPage: number,
-  sortStatusField : string | undefined,
-  sortStatusString : "asc" | "desc" | "",
-}
