+++ ui/src/store/api/api_keys.ts
@@ -1,9 +1,9 @@
 import {
-  ApiKeyValidation, ApiKeyEdit,
-  ApiKeyRemove } from "@/interfaces/IUserLogin";
+  IApiKey, ApiKeyEdit,
+  ApiKeyRemove } from "@/interfaces/IApiKey";
 import { apiKeysApi } from "@/api/http";
 
-export const generateApiKey = async (data: ApiKeyValidation) => apiKeysApi.apiKeyCreate({
+export const generateApiKey = async (data: IApiKey) => apiKeysApi.apiKeyCreate({
   name: data.name,
   role: data.role,
   expires_at: data.expires_at,
@@ -12,20 +12,10 @@ export const generateApiKey = async (data: ApiKeyValidation) => apiKeysApi.apiKe
 export const getApiKey = async (
   page: number,
   perPage: number,
-  sortStatusString : "asc" | "desc" | "",
-  sortStatusField : string | undefined,
-) => {
-  if (sortStatusField && sortStatusString) {
-    return apiKeysApi.apiKeyList(
-      page,
-      perPage,
-      sortStatusString,
-      sortStatusField,
-    );
-  }
+  sortStatusString?: "asc" | "desc",
+  sortStatusField?: string,
+) => apiKeysApi.apiKeyList(page, perPage, sortStatusString, sortStatusField);
 
-  return apiKeysApi.apiKeyList(page, perPage);
-};
 export const removeApiKey = async (data: ApiKeyRemove) => apiKeysApi.apiKeyDelete(data.key);
 
 export const editApiKey = async (data: ApiKeyEdit) => apiKeysApi.apiKeyUpdate(data.key, { name: data.name, role: data.role });
